Greatly modified the general setup of the Plugin.
The main page is now a list of all rounds. Every time the rounds
advance, the last round now opens up for voting, while a new round is
started at the same time.

Once no cards are played in a round, the game pauses, and players will
have to manually draw a new question card until the game continues.